Change hash type to be uintptr_t instead of unsigned int
This fixes warnings on platforms where uintptr_t is an
unsigned long instead of an unsigned int

+++ Foreign/LuaBridge/Source/LuaBridge/RefCountedPtr.h
@@ -51,7 +51,7 @@ struct RefCountedPtrBase
   {
     size_t operator () (const void * const v) const
     {
-      static __gnu_cxx::hash<unsigned int> H;
+      static __gnu_cxx::hash<uintptr_t> H;
       return H(uintptr_t(v));
     }
   };
